requirements/requirements.txt
# We are not pinning patch version for Boto because:
# • it changes very frequently
# • it is reliably compatible
# • it should improve interoperability with AWS services

arrow==1.2.3
attrs==22.1.0
Automat==20.2.0
boto3>=1.17,<1.19
botocore==1.21.41
certifi==2022.9.24
chardet==5.0.0
click==8.1.3
constantly==15.1.0
docker==6.0.0
gitdb==4.0.9
GitPython==3.1.27
hyperlink==21.0.0
idna==3.4
incremental==21.3.0
jmespath==0.10.0
PyHamcrest==2.0.4
python-dateutil==2.8.2
requests==2.28.1
s3transfer==0.5.0
six==1.16.0
smmap==5.0.0
Twisted==21.7.0
urllib3==1.26.12
websocket-client==1.4.1
zope.interface==5.4.0
